# api-docs/swagger/openapi.yaml
...
openapi: 3.0.3
info:
  title: School Management API
  version: 1.0.0
  description: |
    A comprehensive REST API for managing school operations including classes, students, teachers, subjects, lesson schedules, and attendance tracking.
    
    ## Features
    - **Class Management**: Create and manage school classes with grade levels and letters
    - **Student Management**: Handle student enrollment and information
    - **Teacher Management**: Manage teacher profiles and assignments
    - **Subject Management**: Organize school subjects and curriculum
    - **Lesson Scheduling**: Plan and manage weekly lesson schedules
    - **Lesson Logging**: Track actual lessons that occurred
    - **Attendance & Grading**: Record student attendance and grades
    - **Teacher Assignments**: Link teachers to specific subjects
    
    ## Data Model
    The API follows a relational data model where:
    - Students belong to Classes
    - Teachers are assigned to Subjects via TeacherAssignments
    - LessonSchedules define when subjects are taught to classes
    - LessonLogs record actual lessons that happened
    - StudentLessons track individual student participation, grades, and attendance
servers:
  - url: https://d5ds9jru50sk4al4aiaa.svoluuab.apigw.yandexcloud.net
    description: Local development server
...
paths:
...
  /api/v1/students:
    get:
      tags:
        - Students
      summary: List all students
      operationId: listStudents
      parameters:
        - in: query
          name: class_id
          schema:
            type: integer
          description: Filter by class ID
        - in: query
          name: first_name
          schema:
            type: string
          description: Filter by first name (partial match)
        - in: query
          name: last_name
          schema:
            type: string
          description: Filter by last name (partial match)
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        "200":
          description: A list of students
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Student"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Students
      summary: Create a new student
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  "/api/v1/students/{id}":
    get:
      tags:
        - Students
      summary: Get student by ID
      operationId: getStudent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Students
      summary: Update student
      operationId: updateStudent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Students
      summary: Delete student
      operationId: deleteStudent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Student deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/student-lessons:
    get:
      tags:
        - StudentLessons
      summary: List all student lessons
      operationId: listStudentLessons
      parameters:
        - in: query
          name: student_id
          schema:
            type: integer
          description: Filter by student ID
        - in: query
          name: lesson_id
          schema:
            type: integer
          description: Filter by lesson ID
        - in: query
          name: class_id
          schema:
            type: integer
          description: Filter by class ID
        - in: query
          name: attendance_status
          schema:
            type: string
            minLength: 1
            maxLength: 1
            pattern: "^[PALES]$"
          description: Filter by attendance status (P, A, L, E, S)
        - in: query
          name: has_grade
          schema:
            type: boolean
          description: Filter by whether grade is present
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        "200":
          description: A list of student lessons
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/StudentLesson"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - StudentLessons
      summary: Create a student lesson
      operationId: createStudentLesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentLessonInput"
      responses:
        "201":
          description: Student lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentLesson"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  "/api/v1/student-lessons/{id}":
    get:
      tags:
        - StudentLessons
      summary: Get student lesson by ID
      operationId: getStudentLesson
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student lesson found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentLesson"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - StudentLessons
      summary: Update student lesson
      operationId: updateStudentLesson
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentLessonInput"
      responses:
        "200":
          description: Student lesson updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentLesson"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - StudentLessons
      summary: Delete student lesson
      operationId: deleteStudentLesson
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Student lesson deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/lesson-schedules:
    get:
      tags:
        - LessonSchedule
      summary: List all lesson schedules
      operationId: listLessonSchedules
      parameters:
        - in: query
          name: class_id
          schema:
            type: integer
          description: Filter by class ID
        - in: query
          name: teacher_id
          schema:
            type: integer
          description: Filter by teacher ID
        - in: query
          name: subject_id
          schema:
            type: integer
          description: Filter by subject ID
        - in: query
          name: weekday
          schema:
            type: integer
            minimum: 1
            maximum: 7
          description: Filter by weekday (1=Monday, 7=Sunday)
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        "200":
          description: A list of lesson schedules
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/LessonSchedule"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - LessonSchedule
      summary: Create a new lesson schedule
      operationId: createLessonSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonScheduleInput"
      responses:
        "201":
          description: Lesson schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonSchedule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  "/api/v1/lesson-schedules/{id}":
    get:
      tags:
        - LessonSchedule
      summary: Get lesson schedule by ID
      operationId: getLessonSchedule
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson schedule found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonSchedule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - LessonSchedule
      summary: Update lesson schedule
      operationId: updateLessonSchedule
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonScheduleInput"
      responses:
        "200":
          description: Lesson schedule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonSchedule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - LessonSchedule
      summary: Delete lesson schedule
      operationId: deleteLessonSchedule
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Lesson schedule deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the student
        class_id:
          type: integer
          description: ID of the class the student belongs to
        first_name:
          type: string
          minLength: 1
          maxLength: 50
          description: Student's first name
        last_name:
          type: string
          minLength: 1
          maxLength: 50
          description: Student's last name
        patronymic:
          type: string
          maxLength: 50
          description: Student's patronymic (optional)
      required:
        - class_id
        - first_name
        - last_name
      example:
        id: 1
        class_id: 1
        first_name: "Иван"
        last_name: "Иванов"
        patronymic: "Иванович"
...
    LessonSchedule:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the schedule entry
        subject_id:
          type: integer
          description: ID of the subject
        weekday:
          type: integer
          minimum: 1
          maximum: 7
          description: Day of the week (1=Monday, 7=Sunday)
        number:
          type: integer
          minimum: 1
          maximum: 8
          description: Lesson number in the day (1-8)
        class_id:
          type: integer
          description: ID of the class
        teacher_id:
          type: integer
          description: ID of the teacher
      required:
        - subject_id
        - weekday
        - number
        - class_id
        - teacher_id
      example:
        id: 1
        subject_id: 1
        weekday: 1
        number: 1
        class_id: 1
        teacher_id: 1
...
    StudentLesson:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the student lesson record
        student_id:
          type: integer
          description: ID of the student
        lesson_id:
          type: integer
          description: ID of the lesson log
        grade:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
          description: Grade received (1-5, null if not graded)
        attendance_status:
          type: string
          minLength: 1
          maxLength: 1
          pattern: "^[PALES]$"
          description: Attendance status (P=Present, A=Absent, L=Late, E=Excused, S=Sick)
      required:
        - student_id
        - lesson_id
        - attendance_status
      example:
        id: 1
        student_id: 1
        lesson_id: 1
        grade: 5
        attendance_status: "P"
...
    StudentInput:
      type: object
      properties:
        class_id:
          type: integer
          description: ID of the class the student belongs to
        first_name:
          type: string
          minLength: 1
          maxLength: 50
          description: Student's first name
        last_name:
          type: string
          minLength: 1
          maxLength: 50
          description: Student's last name
        patronymic:
          type: string
          maxLength: 50
          description: Student's patronymic (optional)
      required:
        - class_id
        - first_name
        - last_name
...
    LessonScheduleInput:
      type: object
      properties:
        subject_id:
          type: integer
          description: ID of the subject
        weekday:
          type: integer
          minimum: 1
          maximum: 7
          description: Day of the week (1=Monday, 7=Sunday)
        number:
          type: integer
          minimum: 1
          maximum: 8
          description: Lesson number in the day (1-8)
        class_id:
          type: integer
          description: ID of the class
        teacher_id:
          type: integer
          description: ID of the teacher
      required:
        - subject_id
        - weekday
        - number
        - class_id
        - teacher_id
...
    StudentLessonInput:
      type: object
      properties:
        student_id:
          type: integer
          description: ID of the student
        lesson_id:
          type: integer
          description: ID of the lesson log
        grade:
          type: integer
          minimum: 1
          maximum: 5
          description: Grade received (1-5, optional)
        attendance_status:
          type: string
          minLength: 1
          maxLength: 1
          pattern: "^[PALES]$"
          description: Attendance status (P=Present, A=Absent, L=Late, E=Excused, S=Sick)
      required:
        - student_id
        - lesson_id
        - attendance_status
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items
        pages:
          type: integer
          minimum: 0
          description: Total number of pages
      required:
        - page
        - limit
        - total
        - pages
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
      required:
        - error
        - message

  # Reusable Response Definitions
  responses:
    BadRequest:
      description: Bad Request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "BadRequest"
            message: "Invalid input parameters"
            details:
              field: "grade"
              issue: "Must be between 1 and 12"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "NotFound"
            message: "Resource not found"
    UnprocessableEntity:
      description: Unprocessable Entity - Business logic validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "UnprocessableEntity"
            message: "Business logic validation failed"
            details:
              field: "teacher_id"
              issue: "Teacher is already assigned to this subject"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "InternalServerError"
            message: "An unexpected error occurred"